移出模型 HTTP owner 并接入生成 RPC
